+++ src/Microsoft.DotNet.Build.CloudTestTasks/PackageFiles/CloudTest.Helix.targets
@@ -11,7 +11,7 @@
   <UsingTask TaskName="ZipFileCreateFromDependencyLists" AssemblyFile="$(BuildToolsTaskDir)Microsoft.DotNet.Build.CloudTestTasks.dll"/>
 
   <!-- See Documentation\Samples\CloudTest.Helix.targets.sampleproject for a sample project.
-  
+
        Some helpful Helix environment variables: (Use appropriate-for-OS means to access, i.e. %WINDOWS% or $Linux, $OSX )
        **********************************************************************************************************************************
         HELIX_WORKITEM_PAYLOAD      - Execution folder of helix workitem, current directory when using relative path.
@@ -20,7 +20,7 @@
         HELIX_CORRELATION_PAYLOAD   - Correlation payload folder;  root of where all correlation payloads are unzipped.
         HELIX_WORKITEM_FRIENDLYNAME - Friendly name of work item
        **********************************************************************************************************************************
-    
+
        Variables that CloudTest.Helix.Targets cares about:
        **********************************************************************************************************************************
        Required Properties:
@@ -50,9 +50,15 @@
        HelixJobProperties- Must be JSON.  String describing Helix MC-specific metadata.       Default: <Blank>
        HelixArchLabel    - If HelixJobProperties is not set, we'll use this to fill it out    Default: <Blank>
        HelixConfigLabel  - If HelixJobProperties is not set, we'll use this to fill it out    Default: <Blank>
-       SecondaryQueue    - When submitting multi-stage tests, queue for secondary jobs        Default: <Blank>
-       SecondarySasValidHours - SAS Valid time for token provided for multi-stage tests.      Default: 15.0
-	   
+
+       **********************************************************************************************************************************
+       Re-queuing Properties:
+       **********************************************************************************************************************************
+       UseContinuationRunner  - Use runner script that is capable of sending to another queue when finished  Default: False
+       SecondaryQueue         - When submitting multi-stage tests, queue for secondary jobs                  Default: <Blank>
+       SecondarySasValidHours - SAS Valid time for token provided for multi-stage tests.                     Default: 15.0
+       SecondaryPayloadDir    - Folder to make into workitem payload for continued execution                 Default: <Blank>
+
        **********************************************************************************************************************************
        Required ITaskItems:
        **********************************************************************************************************************************
@@ -86,8 +92,8 @@
     <HelixScriptRoot>%HELIX_SCRIPT_ROOT%\</HelixScriptRoot>
     <RunnerScript>%HELIX_CORRELATION_PAYLOAD%\RunnerScripts\scriptrunner\scriptrunner.py</RunnerScript>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(TargetsWindows)' == 'true' AND '$(UseNativeCompileRunner)' == 'true' ">
-    <RunnerScript>%HELIX_CORRELATION_PAYLOAD%\RunnerScripts\scriptrunner\nativecompilerunner.py --secondary_queue $(SecondaryQueue) --secondary_payload_dir %HELIX_WORKITEM_PAYLOAD%\native </RunnerScript>
+  <PropertyGroup Condition="'$(TargetsWindows)' == 'true' AND '$(UseContinuationRunner)' == 'true' ">
+    <RunnerScript>%HELIX_CORRELATION_PAYLOAD%\RunnerScripts\scriptrunner\continuationrunner.py --next_queue $(SecondaryQueue) --next_payload_dir $(SecondaryPayloadDir)</RunnerScript>
     <SecondarySasValidHours Condition="'$(SecondarySasValidHours)' == ''">15.0</SecondarySasValidHours>
   </PropertyGroup>
   <PropertyGroup Condition="'$(TargetsWindows)' != 'true' AND '$(UseScriptRunner)' == 'true' ">
